Implement basic version of game
ComputerPlayer will create code and HumanPlayer will get 12 turns to guess
the code. If HumanPlayer can guess the code in 12 turns he wins. This is
simple version of our game as HumanPlayer will always be guesser of the code
and he can't choose if he wants to be the 'code creator' or 'code breaker'.